refactor: sidebarcomponent moved to the blocks comp folder
